Add files via upload
Libraries:
iostream: This library allows input and output streams, used here for user input and program output.
cstdlib: Provides functions like rand (for generating pseudo-random numbers) and srand (seeding the random number generator).
ctime: Includes functionality related to time and is used here to get the current time, which is then passed to srand for seeding the random number generator.
srand and rand:
srand ((unsigned int) time(NULL)): Seeds the random number generator using the current time. Seeding is crucial to produce different sequences of random numbers each time the program runs.
int Number = (rand() % 100) + 1;: Generates a random number between 1 and 100 (rand() % 100 produces a number between 0 and 99, and adding 1 shifts it to a range between 1 and 100). This is the number the user needs to guess.
Game Logic:
The program prompts the user to enter a guess between 1 and 100.
The input guess is compared to the randomly generated Number.
If the guess is higher than Number, the program prompts the user to guess lower.
If the guess is lower than Number, the program prompts the user to guess higher.
If the guess matches Number, the program declares the user as the winner and ends the game.
do-while Loop:
The game logic is encapsulated within a do-while loop. This loop ensures that the game continues until the user correctly guesses the number (guess != Number).